pyproject.toml
[tool.poetry]
name = "neonotes"
version = "0.1.0"
description = ""
authors = ["serg-e <smosesov@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
google-cloud-vision = "^1.0.0"
pdf2image = "^1.13.1"
notion = "^0.0.25"
jupyter = "^1.0.0"
google-api-python-client = "^1.8.3"
google-auth-httplib2 = "^0.0.3"
google-auth-oauthlib = "^0.4.1"
grpcio = "^1.29.0"
pillow = "^7.1.2"
argparse = "^1.4.0"
fuzzywuzzy = "^0.18.0"
python-Levenshtein = "^0.12.0"
oauth2 = "^1.9.0"
flask-login = "^0.5.0"
flask = "^1.1.2"
oauthlib = "^3.1.0"
pyOpenSSL = "^19.1.0"
sqlalchemy = "^1.3.17"
Flask-SQLAlchemy = "^2.4.3"
flask-migrate = "^2.5.3"
flask-script = "^2.0.6"
psycopg2 = "^2.8.5"
python-dotenv = "^0.13.0"
evernote3 = "^1.25.13"
celery = "^4.4.5"

[tool.poetry.dev-dependencies]
pylint = "^2.5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
